backend/pyproject.toml
[project]
name = "microprelegal-backend"
version = "0.1.0"
description = "FastAPI runtime for the MicroPrelegal V1 foundation."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "litellm>=1.82.6",
  "pydantic>=2.12.5",
  "python-dotenv>=1.2.2",
  "uvicorn>=0.34.0,<1.0.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
